+++ src/module.tsx
@@ -4,6 +4,7 @@ import { AppPlugin } from '@grafana/data';
 // @ts-ignore new API that is not yet in stable release
 import { sidecarServiceSingleton_EXPERIMENTAL } from '@grafana/runtime';
 import pluginJson from './plugin.json';
+import { exposedComponents } from 'exposedComponents';
 import { linkConfigs } from 'utils/links';
 
 const App = lazy(() => import('./components/App/App'));
@@ -16,15 +17,19 @@ export const plugin = new AppPlugin<{}>().setRootPage(App).addConfigPage({
   id: 'configuration',
 })
 .addLink({
-  title: 'traces',
-  description: 'Open traces',
+  title: 'traces drilldown',
+  description: 'Open in Traces Drilldown',
   icon: 'align-left',
   targets: 'grafana-lokiexplore-app/toolbar-open-related/v1',
-  onClick: (e, helpers) => {
-    sidecarServiceSingleton_EXPERIMENTAL?.openApp(pluginJson.id, helpers.context);
+  onClick: () => {
+    sidecarServiceSingleton_EXPERIMENTAL?.openAppV3({ pluginId: pluginJson.id, path: '/explore' });
   },
 });
 
 for (const linkConfig of linkConfigs) {
   plugin.addLink(linkConfig);
+}
+
+for (const exposedComponentConfig of exposedComponents) {
+  plugin.exposeComponent(exposedComponentConfig);
 }