+++ src/utils/shared.ts
@@ -13,6 +13,7 @@ export const PLUGIN_BASE_URL = `/a/${PLUGIN_ID}`;
 export const EXPLORATIONS_ROUTE = `${PLUGIN_BASE_URL}/${ROUTES.Explore}`;
 
 export const DATASOURCE_LS_KEY = 'grafana.explore.traces.datasource';
+export const HOMEPAGE_FILTERS_LS_KEY = 'grafana.explore.traces.homepage.filters';
 
 export const GRID_TEMPLATE_COLUMNS = 'repeat(auto-fit, minmax(400px, 1fr))';
 
@@ -21,9 +22,12 @@ export const EMPTY_STATE_ERROR_REMEDY_MESSAGE = 'Please try removing some filter
 
 export const VAR_DATASOURCE = 'ds';
 export const VAR_DATASOURCE_EXPR = '${ds}';
+export const VAR_PRIMARY_SIGNAL = 'primarySignal';
 export const VAR_FILTERS = 'filters';
-export const VAR_FILTERS_EXPR = '${filters}';
+export const VAR_FILTERS_EXPR = '${primarySignal} && ${filters}';
+export const VAR_HOME_FILTER = 'homeFilter';
 export const VAR_GROUPBY = 'groupBy';
+export const VAR_SPAN_LIST_COLUMNS = 'spanListColumns';
 export const VAR_METRIC = 'metric';
 export const VAR_LATENCY_THRESHOLD = 'latencyThreshold';
 export const VAR_LATENCY_THRESHOLD_EXPR = '${latencyThreshold}';
@@ -36,6 +40,8 @@ export const RESOURCE = 'Resource';
 export const SPAN = 'Span';
 export const RESOURCE_ATTR = 'resource.';
 export const SPAN_ATTR = 'span.';
+export const EVENT_ATTR = 'event.';
+export const EVENT_INTRINSIC = 'event:';
 
 export const radioAttributesResource = [
   // https://opentelemetry.io/docs/specs/semconv/resource/
@@ -74,6 +80,16 @@ export const ignoredAttributes = [
   'trace:duration',
   'trace:id',
   'traceDuration',
+];
+export const ignoredAttributesHomeFilter = [
+  'status',
+  'span:status',
+  'rootName',
+  'rootService',
+  'rootServiceName',
+  'trace:rootName',
+  'trace:rootService',
+  'trace:rootServiceName',
 ];
 // Limit maximum options in select dropdowns for performance reasons
 export const maxOptions = 1000;