+++ src/utils/utils.ts
@@ -3,6 +3,7 @@ import {
   AdHocFiltersVariable,
   CustomVariable,
   DataSourceVariable,
+  SceneDataQuery,
   SceneDataState,
   sceneGraph,
   SceneObject,
@@ -18,15 +19,18 @@ import {
   VAR_DATASOURCE_EXPR,
   VAR_FILTERS,
   VAR_GROUPBY,
+  VAR_HOME_FILTER,
   VAR_LATENCY_PARTIAL_THRESHOLD,
   VAR_LATENCY_THRESHOLD,
   VAR_METRIC,
+  VAR_PRIMARY_SIGNAL,
+  VAR_SPAN_LIST_COLUMNS,
 } from './shared';
-import { primarySignalOptions } from '../pages/Explore/primary-signals';
 import { TracesByServiceScene } from 'components/Explore/TracesByService/TracesByServiceScene';
 import { ActionViewType } from '../components/Explore/TracesByService/Tabs/TabsBarScene';
 import { LocationService } from '@grafana/runtime';
 import { Home } from 'pages/Home/Home';
+import { PrimarySignalVariable } from 'pages/Explore/PrimarySignalVariable';
 
 export function getTraceExplorationScene(model: SceneObject): TraceExploration {
   return sceneGraph.getAncestor(model, TraceExploration);
@@ -47,15 +51,16 @@ export function newTracesExploration(
 ): TraceExploration {
   return new TraceExploration({
     initialDS,
-    initialFilters: initialFilters ?? [primarySignalOptions[0].filter],
+    initialFilters: initialFilters ?? [],
     $timeRange: new SceneTimeRange({ from: 'now-30m', to: 'now' }),
     locationService,
   });
 }
 
-export function newHome(initialDS?: string): Home {
+export function newHome(initialFilters: AdHocVariableFilter[], initialDS?: string): Home {
   return new Home({
     initialDS,
+    initialFilters,
     $timeRange: new SceneTimeRange({ from: 'now-30m', to: 'now' }),
   });
 }
@@ -65,7 +70,7 @@ export function getErrorMessage(data: SceneDataState) {
 }
 
 export function getNoDataMessage(context: string) {
-  return `No data for selected data source. Select another to see ${context}.`;
+  return `No data for selected data source and filter. Select another to see ${context}.`;
 }
 
 export function getUrlForExploration(exploration: TraceExploration) {
@@ -89,7 +94,7 @@ export function getAttributesAsOptions(attributes: string[]) {
   return attributes.map((attribute) => ({ label: attribute, value: attribute }));
 }
 
-export function getLabelValue(frame: DataFrame, labelName?: string) {
+export function getLabelKey(frame: DataFrame) {
   const labels = frame.fields.find((f) => f.type === 'number')?.labels;
 
   if (!labels) {
@@ -101,6 +106,21 @@ export function getLabelValue(frame: DataFrame, labelName?: string) {
     return 'No labels';
   }
 
+  return keys[0].replace(/"/g, '');
+}
+
+export function getLabelValue(frame: DataFrame, labelName?: string) {
+  const labels = frame.fields.find((f) => f.type === 'number')?.labels;
+
+  if (!labels) {
+    return 'No labels';
+  }
+
+  const keys = Object.keys(labels).filter((k) => k !== 'p'); // remove the percentile label
+  if (keys.length === 0) {
+    return 'No labels';
+  }
+
   return labels[labelName || keys[0]].replace(/"/g, '');
 }
 
@@ -108,6 +128,14 @@ export function getGroupByVariable(scene: SceneObject): CustomVariable {
   const variable = sceneGraph.lookupVariable(VAR_GROUPBY, scene);
   if (!(variable instanceof CustomVariable)) {
     throw new Error('Group by variable not found');
+  }
+  return variable;
+}
+
+export function getSpanListColumnsVariable(scene: SceneObject): CustomVariable {
+  const variable = sceneGraph.lookupVariable(VAR_SPAN_LIST_COLUMNS, scene);
+  if (!(variable instanceof CustomVariable)) {
+    throw new Error('Span list columns variable not found');
   }
   return variable;
 }
@@ -144,12 +172,34 @@ export function getFiltersVariable(scene: SceneObject): AdHocFiltersVariable {
   return variable;
 }
 
+export function getPrimarySignalVariable(scene: SceneObject): PrimarySignalVariable {
+  const variable = sceneGraph.lookupVariable(VAR_PRIMARY_SIGNAL, scene);
+  if (!(variable instanceof PrimarySignalVariable)) {
+    throw new Error('Primary signal variable not found');
+  }
+  return variable;
+}
+
+export function getHomeFilterVariable(scene: SceneObject): AdHocFiltersVariable {
+  const variable = sceneGraph.lookupVariable(VAR_HOME_FILTER, scene);
+  if (!(variable instanceof AdHocFiltersVariable)) {
+    throw new Error('Home filter variable not found');
+  }
+  return variable;
+}
+
 export function getDatasourceVariable(scene: SceneObject): DataSourceVariable {
   const variable = sceneGraph.lookupVariable(VAR_DATASOURCE, scene);
   if (!(variable instanceof DataSourceVariable)) {
     throw new Error('Datasource variable not found');
   }
   return variable;
+}
+
+export function getCurrentStep(scene: SceneObject): number | undefined {
+  const data = sceneGraph.getData(scene).state.data;
+  const targetQuery = data?.request?.targets[0];
+  return targetQuery ? (targetQuery as SceneDataQuery).step : undefined;
 }
 
 export function shouldShowSelection(tab?: ActionViewType): boolean {
@@ -163,3 +213,12 @@ export function getMetricValue(scene: SceneObject) {
 export function fieldHasEmptyValues(data: SceneDataState) {
   return data?.data?.series[0].fields?.some((v) => v.values.every((e) => e === undefined)) ?? false;
 }
+
+export const isNumber = /^-?\d+\.?\d*$/;
+
+export const formatLabelValue = (value: string) => {
+  if (!isNumber.test(value) && typeof value === 'string' && !value.startsWith('"') && !value.endsWith('"')) {
+    return `"${value}"`;
+  }
+  return value;
+};